Se agregaron cambios a la creacion de anticipo, mejoras, ahora la moneda de BS se multiplica por la tasa, o se divide si es el monto en DS. Ahora ya hay bien definida una clase para TasaUSD, falta ponerle las modificaciones y guardados de datos a la tabla de anticipos

+++ src/clases/anticipos/ModificarAnticipo.java
@@ -363,8 +363,8 @@ public class ModificarAnticipo extends javax.swing.JFrame {
         if (!(Codigotxt.getText().equals("")) && !(IDtxt.getText().equals("")) && !(RazonSocialtxt.getText().equals("")) && !(NroAnticipotxt.getText().isEmpty())) {
             String motivo_anticipo = MotivoAnticipoCB.getSelectedItem().toString();
             String fecha = Fechatxt.getText();
-            float monto_bs = Float.parseFloat(MontoBStxt.getText());
-            float monto_ds = Float.parseFloat(MontoDStxt.getText());
+            Double monto_bs = Double.parseDouble(MontoBStxt.getText());
+            Double monto_ds = Double.parseDouble(MontoDStxt.getText());
             String aprobacion = Aprobaciontxt.getText();
             String descontarODP = DescontarCB.getSelectedItem().toString();
             String observaciones = Observaciontxt.getText();
